Towards strongly aniso nonhomo systems.
